pyproject.toml
[tool.poetry]
name = "az"
description = "Vodka A-Z"
version = "1.0.0"
authors = ["OpenAlcoholics <team@openalcoholics.group>"]
license = "MIT"

[tool.poetry.dependencies]
aiohttp = "^3.0.0"
certifi = "2022.12.7"
click = "^8.1.3"
openai = "^0.27.0"
python = "~3.11"
python-dotenv = "^1.0.0"
python-telegram-bot = "20.1"
sentry-sdk = "^1.14.0"

[tool.poetry.group.dev.dependencies]
black = "^23.0.0"
mypy = "~1.1"
types-requests = "^2.28.11"

[build-system]
requires = ["poetry-core>=1.3.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
strict = true
warn_return_any = false

[[tool.mypy.overrides]]
module = "tests.*"
strict = false
check_untyped_defs = true
disallow_untyped_defs = false
